refactor: Split Paperlist component
- `PaperListWrapper` fetch/get data & format it.
- `PaperListByContact` loops over an array of contacts
with their respective papers.
- `PaperList` loops over all the papers of each contact.
