+++ src/components/Papers/PapersListByContact.jsx
@@ -1,0 +1,35 @@
+import React, { Fragment } from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import PropTypes from 'prop-types'
+
+import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
+import List from 'cozy-ui/transpiled/react/MuiCozyTheme/List'
+import ListSubheader from 'cozy-ui/transpiled/react/MuiCozyTheme/ListSubheader'
+
+import PapersList from 'src/components/Papers/PapersList'
+
+const useStyles = makeStyles({
+  root: { textIndent: '1rem' }
+})
+
+const PapersListByContact = ({ paperslistByContact }) => {
+  const classes = useStyles()
+  const { isMobile } = useBreakpoints()
+
+  return (
+    <List>
+      {paperslistByContact.map(({ contact, papers }, idx) => (
+        <Fragment key={idx}>
+          <ListSubheader classes={isMobile && classes}>{contact}</ListSubheader>
+          <PapersList papers={papers} />
+        </Fragment>
+      ))}
+    </List>
+  )
+}
+
+PapersListByContact.propTypes = {
+  paperslistByContact: PropTypes.arrayOf(PropTypes.object)
+}
+
+export default PapersListByContact
